Version 0.0.2 Updates
This version is an updated, refactored engine that contains the following changes:
* Attempts to utilize C++11 & C++14 constructs.
* Rework so that the engine invokes the OS and Graphics device rather than the OS invoking everything else.
* Refactoring of a majority of the code to hopefully, properly use the const keyword on getter methods and implementing constant pointers where necessary.
* Upgraded some of the DX11 constructs and included the WICTextureLoader.h/.cpp from the Microsoft github account to load textures.
* Added more logic to get OpenGL running to some extent.

+++ CoffeeEngine/Include/Utility/LoggingTypes.h
@@ -50,7 +50,7 @@ namespace CoffeeEngine
 					"Diagnostic"
 				};
 
-				auto logLevelIntegral = PrecomputeLog2Byte(static_cast<unsigned>(logLevelType));
+				auto logLevelIntegral = PrecomputeLog2Byte(static_cast<uint16_t>(logLevelType));
 				return (logLevelIntegral < std::size(LogLevelTypeNames)) ? LogLevelTypeNames[logLevelIntegral] : "Undefined";
 			}
